+++ common/coconut-interface/src/error.rs
@@ -1,6 +1,7 @@
 // Copyright 2021 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 
+use nymcoconut::CoconutError;
 use thiserror::Error;
 
 #[derive(Debug, Error)]
@@ -11,9 +12,6 @@ pub enum CoconutInterfaceError {
     #[error("Could not decode base 58 string - {0}")]
     MalformedString(#[from] bs58::decode::Error),
 
-    #[error("Not enough public attributes were specified")]
-    NotEnoughPublicAttributes,
-
-    #[error("Could not recover bandwidth value")]
-    InvalidBandwidth,
+    #[error("Coconut error - {0}")]
+    CoconutError(#[from] CoconutError),
 }