+++ common/network-defaults/src/all.rs
@@ -52,6 +52,14 @@ impl Network {
         self.details().bandwidth_claim_contract_address
     }
 
+    pub fn coconut_bandwidth_contract_address(&self) -> &str {
+        self.details().coconut_bandwidth_contract_address
+    }
+
+    pub fn multisig_contract_address(&self) -> &str {
+        self.details().multisig_contract_address
+    }
+
     pub fn rewarding_validator_address(&self) -> &str {
         self.details().rewarding_validator_address
     }