+++ nym-wallet/src-tauri/src/operations/simulate/mixnet.rs
@@ -20,7 +20,7 @@ pub async fn simulate_bond_gateway(
     let pledge = pledge.into_backend_coin(guard.current_network().denom())?;
 
     let client = guard.current_client()?;
-    let mixnet_contract = client.nymd.mixnet_contract_address()?;
+    let mixnet_contract = client.nymd.mixnet_contract_address();
     let gas_price = client.nymd.gas_price().clone();
 
     // TODO: I'm still not 100% convinced whether this should be exposed here or handled somewhere else in the client code
@@ -44,7 +44,7 @@ pub async fn simulate_unbond_gateway(
     let guard = state.read().await;
 
     let client = guard.current_client()?;
-    let mixnet_contract = client.nymd.mixnet_contract_address()?;
+    let mixnet_contract = client.nymd.mixnet_contract_address();
     let gas_price = client.nymd.gas_price().clone();
 
     let msg = client.nymd.wrap_contract_execute_message(
@@ -68,7 +68,7 @@ pub async fn simulate_bond_mixnode(
     let pledge = pledge.into_backend_coin(guard.current_network().denom())?;
 
     let client = guard.current_client()?;
-    let mixnet_contract = client.nymd.mixnet_contract_address()?;
+    let mixnet_contract = client.nymd.mixnet_contract_address();
     let gas_price = client.nymd.gas_price().clone();
 
     let msg = client.nymd.wrap_contract_execute_message(
@@ -91,7 +91,7 @@ pub async fn simulate_unbond_mixnode(
     let guard = state.read().await;
 
     let client = guard.current_client()?;
-    let mixnet_contract = client.nymd.mixnet_contract_address()?;
+    let mixnet_contract = client.nymd.mixnet_contract_address();
     let gas_price = client.nymd.gas_price().clone();
 
     let msg = client.nymd.wrap_contract_execute_message(
@@ -112,7 +112,7 @@ pub async fn simulate_update_mixnode(
     let guard = state.read().await;
 
     let client = guard.current_client()?;
-    let mixnet_contract = client.nymd.mixnet_contract_address()?;
+    let mixnet_contract = client.nymd.mixnet_contract_address();
     let gas_price = client.nymd.gas_price().clone();
 
     let msg = client.nymd.wrap_contract_execute_message(
@@ -137,7 +137,7 @@ pub async fn simulate_delegate_to_mixnode(
     let delegation = amount.into_backend_coin(guard.current_network().denom())?;
 
     let client = guard.current_client()?;
-    let mixnet_contract = client.nymd.mixnet_contract_address()?;
+    let mixnet_contract = client.nymd.mixnet_contract_address();
     let gas_price = client.nymd.gas_price().clone();
 
     let msg = client.nymd.wrap_contract_execute_message(
@@ -160,7 +160,7 @@ pub async fn simulate_undelegate_from_mixnode(
     let guard = state.read().await;
 
     let client = guard.current_client()?;
-    let mixnet_contract = client.nymd.mixnet_contract_address()?;
+    let mixnet_contract = client.nymd.mixnet_contract_address();
     let gas_price = client.nymd.gas_price().clone();
 
     let msg = client.nymd.wrap_contract_execute_message(