+++ nym-wallet/src-tauri/src/operations/simulate/vesting.rs
@@ -21,7 +21,7 @@ pub async fn simulate_vesting_bond_gateway(
     let pledge = pledge.into_backend_coin(guard.current_network().denom())?;
 
     let client = guard.current_client()?;
-    let vesting_contract = client.nymd.vesting_contract_address()?;
+    let vesting_contract = client.nymd.vesting_contract_address();
     let gas_price = client.nymd.gas_price().clone();
 
     let msg = client.nymd.wrap_contract_execute_message(
@@ -45,7 +45,7 @@ pub async fn simulate_vesting_unbond_gateway(
     let guard = state.read().await;
 
     let client = guard.current_client()?;
-    let vesting_contract = client.nymd.vesting_contract_address()?;
+    let vesting_contract = client.nymd.vesting_contract_address();
     let gas_price = client.nymd.gas_price().clone();
 
     let msg = client.nymd.wrap_contract_execute_message(
@@ -69,7 +69,7 @@ pub async fn simulate_vesting_bond_mixnode(
     let pledge = pledge.into_backend_coin(guard.current_network().denom())?;
 
     let client = guard.current_client()?;
-    let vesting_contract = client.nymd.vesting_contract_address()?;
+    let vesting_contract = client.nymd.vesting_contract_address();
     let gas_price = client.nymd.gas_price().clone();
 
     let msg = client.nymd.wrap_contract_execute_message(
@@ -93,7 +93,7 @@ pub async fn simulate_vesting_unbond_mixnode(
     let guard = state.read().await;
 
     let client = guard.current_client()?;
-    let vesting_contract = client.nymd.vesting_contract_address()?;
+    let vesting_contract = client.nymd.vesting_contract_address();
     let gas_price = client.nymd.gas_price().clone();
 
     let msg = client.nymd.wrap_contract_execute_message(
@@ -114,7 +114,7 @@ pub async fn simulate_vesting_update_mixnode(
     let guard = state.read().await;
 
     let client = guard.current_client()?;
-    let vesting_contract = client.nymd.vesting_contract_address()?;
+    let vesting_contract = client.nymd.vesting_contract_address();
     let gas_price = client.nymd.gas_price().clone();
 
     let msg = client.nymd.wrap_contract_execute_message(
@@ -138,7 +138,7 @@ pub async fn simulate_withdraw_vested_coins(
     let amount = amount.into_backend_coin(guard.current_network().denom())?;
 
     let client = guard.current_client()?;
-    let vesting_contract = client.nymd.vesting_contract_address()?;
+    let vesting_contract = client.nymd.vesting_contract_address();
     let gas_price = client.nymd.gas_price().clone();
 
     let msg = client.nymd.wrap_contract_execute_message(