+++ validator-api/src/rewarded_set_updater/error.rs
@@ -8,9 +8,6 @@ use validator_client::ValidatorClientError;
 
 #[derive(Debug, Error)]
 pub enum RewardingError {
-    #[error("Could not distribute rewards as the contract address was unspecified")]
-    UnspecifiedContractAddress,
-
     // #[error("There were no mixnodes to reward (network is dead)")]
     // NoMixnodesToReward,
     #[error("Failed to execute the smart contract - {0}")]